# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /course/weeks/{weekNumber}:
    get:
      operationId: getWeek
      tags: [course]
      summary: Get one week's lectures and assignment list
      parameters:
        - name: weekNumber
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 4 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
  /course/lectures/{lectureId}:
    get:
      operationId: getLecture
      tags: [course]
      summary: Get the full lecture text for a topic
      parameters:
        - name: lectureId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
  /course/lectures/{lectureId}/rewrite:
    post:
      operationId: rewriteLecture
      tags: [course]
      summary: Rewrite the lecture from the student's own instruction
      description: >-
        Produces a reader-directed rewrite of the lecture (e.g. more examples on
        a point, a clearer illustration of a principle, shorter sentences) while
        preserving every concept and learning objective. Persists the result as
        the lecture's custom version and returns the updated lecture.
      parameters:
        - name: lectureId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewriteLectureInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    delete:
      operationId: clearLectureRewrite
      tags: [course]
      summary: Discard the custom rewrite and revert to the original lecture
      parameters:
        - name: lectureId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
...
components:
  schemas:
...
    Lecture:
      type: object
      required: [id, topicId, title, weekNumber, body]
      properties:
        id: { type: integer }
        topicId: { type: integer }
        title: { type: string }
        weekNumber: { type: integer }
        body:
          type: string
          description: Short Markdown lecture text (the baseline / minimum-detail version). The frontend lets users select passages and send them to the tutor.
        bodyMedium:
          type: ["string", "null"]
          description: Medium-length version with more explanation and more examples. Null if not yet generated.
        bodyLong:
          type: ["string", "null"]
          description: Long version with the most explanation and the most examples. Null if not yet generated.
        bodyCustom:
          type: ["string", "null"]
          description: A reader-directed rewrite of the lecture produced from the student's own instruction (e.g. add more examples, illustrate a principle, shorter sentences). Null until the student requests a rewrite.
        customInstruction:
          type: ["string", "null"]
          description: The most recent instruction the student gave to produce bodyCustom. Null when there is no custom rewrite.
    Week:
      type: object
      required: [weekNumber, title, lectures, assignments]
      properties:
        weekNumber: { type: integer }
        title: { type: string }
        summary: { type: ["string", "null"] }
        lectures:
          type: array
          items:
            $ref: "#/components/schemas/LectureRef"
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentSummary"
    LectureRef:
      type: object
      required: [id, title, topicId]
      properties:
        id: { type: integer }
        title: { type: string }
        topicId: { type: integer }
    AssignmentSummary:
      type: object
      required: [id, kind, title, weekNumber, problemCount, isTimed, status]
      properties:
        id: { type: integer }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        title: { type: string }
        weekNumber: { type: integer }
        problemCount: { type: integer }
        isTimed: { type: boolean }
        timeLimitMinutes: { type: ["integer", "null"] }
        status:
          type: string
          enum: [not_started, in_progress, submitted]
        bestScore: { type: ["number", "null"] }
        lastAttemptId: { type: ["integer", "null"] }
...
    RewriteLectureInput:
      type: object
      required: [instruction]
      properties:
        instruction:
          type: string
          minLength: 1
          maxLength: 1000
          description: The student's plain-language instruction for how to rewrite the lecture (e.g. "add more examples to the section on intrinsic goodness", "use shorter sentences", "illustrate the open-question argument with a concrete case").
        baseLevel:
          type: string
          enum: [short, medium, long, custom]
          description: Which version of the lecture to rewrite from. Defaults to "short". Use "custom" to refine an existing rewrite further.
...