# lib/api-zod/src/generated/types/lecture.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lecture {
  id: number;
  topicId: number;
  title: string;
  weekNumber: number;
  /** Short Markdown lecture text (the baseline / minimum-detail version). The frontend lets users select passages and send them to the tutor. */
  body: string;
  /**
     * Medium-length version with more explanation and more examples. Null if not yet generated.
     * @nullable
     */
  bodyMedium?: string | null;
  /**
     * Long version with the most explanation and the most examples. Null if not yet generated.
     * @nullable
     */
  bodyLong?: string | null;
  /**
     * A reader-directed rewrite of the lecture produced from the student's own instruction (e.g. add more examples, illustrate a principle, shorter sentences). Null until the student requests a rewrite.
     * @nullable
     */
  bodyCustom?: string | null;
  /**
     * The most recent instruction the student gave to produce bodyCustom. Null when there is no custom rewrite.
     * @nullable
     */
  customInstruction?: string | null;
}