# lib/api-zod/src/generated/types/rewriteLectureInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RewriteLectureInputBaseLevel } from './rewriteLectureInputBaseLevel';

export interface RewriteLectureInput {
  /**
     * The student's plain-language instruction for how to rewrite the lecture (e.g. "add more examples to the section on intrinsic goodness", "use shorter sentences", "illustrate the open-question argument with a concrete case").
     * @minLength 1
     * @maxLength 1000
     */
  instruction: string;
  /** Which version of the lecture to rewrite from. Defaults to "short". Use "custom" to refine an existing rewrite further. */
  baseLevel?: RewriteLectureInputBaseLevel;
}